requirements.txt
Pillow
matplotlib
streamlit
pandas
numpy
POT

